feat: 환율데이터 db 저장 및 조회 api 추가
