+++ src/main/java/com/project/korex/exchangeRate/repository/EximExchangeRateRepository.java
@@ -1,7 +1,6 @@
 package com.project.korex.exchangeRate.repository;
 
-import com.project.korex.exchangeRate.entity.CurrencyRate;
-import com.project.korex.exchangeRate.entity.ExchangeRate;
+import com.project.korex.exchangeRate.entity.EximExchangeRate;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,9 +8,12 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface CurrencyRateRepository extends JpaRepository<CurrencyRate, Long> {
+public interface EximExchangeRateRepository extends JpaRepository<EximExchangeRate, Long> {
 
+    // 날짜 기준 삭제
     void deleteByBaseDate(LocalDate baseDate);
 
-    List<ExchangeRate> findByBaseDate(LocalDate baseDate);
+    // 날짜별 조회
+    List<EximExchangeRate> findByBaseDate(LocalDate baseDate);
+
 }
